+++ source-code-git/fs/src/index.ts
@@ -1,4 +1,2 @@
-export type { Filesystem, FileData, FilesystemError } from "./schema.js"
-export { createMemoryFs } from "./createMemoryFs.js"
-export { fromNodeFs } from "./fromNodeFs.js"
-export { toNodeFs } from "./toNodeFs.js"
+export type { Filesystem } from "./interface.js"
+export { createMemoryFs } from "./implementations/memoryFs.js"
