+++ source-code-git/fs/src/interface.ts
