--- source-code-git/fs/src_new/index.ts
@@ -1,2 +1,0 @@
-export type { Filesystem } from "./interface.js"
-export { createMemoryFs } from "./implementations/memoryFs.js"
