+++ source-code-git/fs/src_old/index.ts
@@ -1,0 +1,4 @@
+export type { Filesystem, FileData, FilesystemError } from "./schema.js"
+export { createMemoryFs } from "./createMemoryFs.js"
+export { fromNodeFs } from "./fromNodeFs.js"
+export { toNodeFs } from "./toNodeFs.js"
